+++ packages/abi/lib.esm/coders/abstract-coder.js
@@ -1,5 +1,5 @@
 "use strict";
-import { arrayify, concat, hexlify } from "@ethersproject/bytes";
+import { arrayify, concat, hexConcat, hexlify } from "@ethersproject/bytes";
 import { BigNumber } from "@ethersproject/bignumber";
 import { defineReadOnly } from "@ethersproject/properties";
 import { Logger } from "@ethersproject/logger";
@@ -41,20 +41,28 @@ export class Coder {
 export class Writer {
     constructor(wordSize) {
         defineReadOnly(this, "wordSize", wordSize || 32);
-        this._data = arrayify([]);
+        this._data = [];
+        this._dataLength = 0;
         this._padding = new Uint8Array(wordSize);
     }
-    get data() { return hexlify(this._data); }
-    get length() { return this._data.length; }
+    get data() {
+        return hexConcat(this._data);
+    }
+    get length() { return this._dataLength; }
     _writeData(data) {
-        this._data = concat([this._data, data]);
+        this._data.push(data);
+        this._dataLength += data.length;
         return data.length;
+    }
+    appendWriter(writer) {
+        return this._writeData(concat(writer._data));
     }
     // Arrayish items; padded on the right to wordSize
     writeBytes(value) {
         let bytes = arrayify(value);
-        if (bytes.length % this.wordSize) {
-            bytes = concat([bytes, this._padding.slice(bytes.length % this.wordSize)]);
+        const paddingOffset = bytes.length % this.wordSize;
+        if (paddingOffset) {
+            bytes = concat([bytes, this._padding.slice(paddingOffset)]);
         }
         return this._writeData(bytes);
     }
@@ -76,10 +84,11 @@ export class Writer {
         return this._writeData(this._getValue(value));
     }
     writeUpdatableValue() {
-        let offset = this.length;
-        this.writeValue(0);
+        const offset = this._data.length;
+        this._data.push(this._padding);
+        this._dataLength += this.wordSize;
         return (value) => {
-            this._data.set(this._getValue(value), offset);
+            this._data[offset] = this._getValue(value);
         };
     }
 }