+++ packages/hardware-wallets/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "hardware-wallets/5.0.7";
+export declare const version = "hardware-wallets/5.0.8";