+++ packages/providers/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "providers/5.0.10";
+export declare const version = "providers/5.0.11";